requirements.render.txt
# Minimal Requirements for Render Free Tier (512MB limit)
# This version excludes heavy OCR and PDF generation dependencies

# Core AI
google-generativeai>=0.8.0

# Environment variables
python-dotenv==1.0.0

# PDF text extraction (lightweight)
PyPDF2==3.0.0
PyMuPDF==1.23.0

# API server
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Note: Excluded to avoid C compiler requirements and save memory:
# - reportlab (requires pycairo, needs gcc) - PDF generation disabled
# - Pillow (not needed for text extraction)
# - OCR features (paddleocr, torch, opencv)
# 
# The code gracefully handles missing reportlab by disabling PDF generation.
# Resume analysis and enhancement still work without it.
# 
# For full features including PDF generation, deploy to:
# - Hugging Face Spaces (FREE with 16GB RAM)
# - Render paid tier with build support
